pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "lxml>=6.0.2",
    "openai>=2.15.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
]
